Refactor Battle so it can rerender when switching characters or starting a new battle
